refactor(search): extract reusable SearchView; use it on home and profile pages with by:<npub> prefill
